requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./scripts/pip-compile.sh
#
alembic==1.4.3
    # via
    #   dataset
    #   flask-migrate
aniso8601==8.0.0
    # via flask-restx
attrs==20.3.0
    # via jsonschema
bcrypt==3.1.7
    # via -r requirements.in
boto3==1.13.9
    # via -r requirements.in
botocore==1.16.26
    # via
    #   boto3
    #   s3transfer
certifi==2020.11.8
    # via requests
cffi==1.14.3
    # via
    #   bcrypt
    #   cmarkgfm
    #   pybluemonday
chardet==3.0.4
    # via requests
click==7.1.2
    # via flask
cmarkgfm==0.4.2
    # via -r requirements.in
dataset==1.3.1
    # via -r requirements.in
docutils==0.15.2
    # via botocore
flask==1.1.2
    # via
    #   -r requirements.in
    #   flask-caching
    #   flask-marshmallow
    #   flask-migrate
    #   flask-restx
    #   flask-script
    #   flask-sqlalchemy
flask-caching==1.8.0
    # via -r requirements.in
flask-marshmallow==0.10.1
    # via -r requirements.in
flask-migrate==2.5.3
    # via -r requirements.in
flask-restx==0.5.1
    # via -r requirements.in
flask-script==2.0.6
    # via -r requirements.in
flask-sqlalchemy==2.4.3
    # via
    #   -r requirements.in
    #   flask-migrate
gevent==20.9.0
    # via -r requirements.in
greenlet==0.4.17
    # via gevent
gunicorn==20.0.4
    # via -r requirements.in
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via
    #   -r requirements.in
    #   flask
jinja2==2.11.3
    # via
    #   -r requirements.in
    #   flask
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via flask-restx
mako==1.1.3
    # via alembic
markupsafe==1.1.1
    # via
    #   jinja2
    #   mako
    #   wtforms
marshmallow==2.20.2
    # via
    #   -r requirements.in
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.17.0
    # via -r requirements.in
maxminddb==1.5.4
    # via
    #   -r requirements.in
    #   python-geoacumen-city
passlib==1.7.2
    # via -r requirements.in
pillow==9.0.0
    # via -r requirements.in
pybluemonday==0.0.8
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydantic==1.6.2
    # via -r requirements.in
pymysql==0.9.3
    # via -r requirements.in
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   alembic
    #   botocore
python-dotenv==0.13.0
    # via -r requirements.in
python-editor==1.0.4
    # via alembic
python-geoacumen-city==0.0.1
    # via -r requirements.in
pytz==2020.4
    # via flask-restx
redis==3.5.2
    # via -r requirements.in
requests==2.23.0
    # via -r requirements.in
s3transfer==0.3.3
    # via boto3
six==1.15.0
    # via
    #   bcrypt
    #   flask-marshmallow
    #   flask-restx
    #   jsonschema
    #   python-dateutil
    #   sqlalchemy-utils
    #   tenacity
sqlalchemy==1.3.17
    # via
    #   -r requirements.in
    #   alembic
    #   dataset
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.36.6
    # via -r requirements.in
tenacity==6.2.0
    # via -r requirements.in
urllib3==1.25.11
    # via
    #   botocore
    #   requests
werkzeug==1.0.1
    # via
    #   -r requirements.in
    #   flask
    #   flask-restx
wtforms==2.3.1
    # via -r requirements.in
zope.event==4.5.0
    # via gevent
zope.interface==5.2.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
